Refactor parser to split private and public interfaces
