--- src/parser/mod.rs
@@ -1,4 +1,0 @@
-pub mod aircraft;
-pub mod builder;
-pub mod constants;
-pub mod types;
